requirements.txt
asgiref==3.8.1
Django==5.0.6
sqlparse==0.5.0
tzdata==2024.1
djangorestframework==3.14.0
bleak==0.22.2
